requirements.txt
blessed==1.20.0
blinker==1.7.0
botocore==1.34.49
certifi==2024.2.2
chalice==1.30.0
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
idna==3.6
inquirer==2.10.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
PyMySQL==1.1.0
pyserial==3.5
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0.1
readchar==4.0.5
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==3.0.1
